+++ DatabaseModel/Postar.cs
@@ -21,7 +21,7 @@ namespace DatabaseModel
         }
     
         public int JMBG_Radnika { get; set; }
-        public int PostanskiBroj { get; set; }
+        public decimal PostanskiBroj { get; set; }
         public string DeoGrada { get; set; }
     
         public virtual Radnik Radnik { get; set; }
